Abstract all devices out of the driver
